requirements.txt
SQLAlchemy==0.8.2
Unidecode==0.04.13
feedparser==5.1.3
oauthlib==0.5.0
psycopg2==2.5.1
python-dateutil==2.1
requests==1.2.3
requests-oauthlib==0.3.2
twython==3.0.0
flickrapi==1.4.2
python-instagram==0.8.0
